serialization has been moved to working with files
