+++ AsymmetricEncryption.py
@@ -23,28 +23,6 @@ class AsymmetricEncryption:
             raise RuntimeError(f"Ошибка генерации ключей RSA: {str(e)}")
 
     @staticmethod
-    def serialize_keys(public_key: rsa.RSAPublicKey, public_key_path: str,
-                       private_key: RSAPrivateKey, private_key_path: str) -> None:
-        """Serializing keys to files"""
-        try:
-            with open(public_key_path, 'wb') as f:
-                f.write(public_key.public_bytes(
-                    encoding=serialization.Encoding.PEM,
-                    format=serialization.PublicFormat.SubjectPublicKeyInfo
-                ))
-            print(f"Открытый ключ сохранен в {public_key_path}")
-
-            with open(private_key_path, 'wb') as f:
-                f.write(private_key.private_bytes(
-                    encoding=serialization.Encoding.PEM,
-                    format=serialization.PrivateFormat.PKCS8,
-                    encryption_algorithm=serialization.NoEncryption()
-                ))
-            print(f"Закрытый ключ сохранен в {private_key_path}")
-        except IOError as e:
-            raise RuntimeError(f"Ошибка записи ключей в файл: {str(e)}")
-
-    @staticmethod
     def encrypt_symmetric_key(symmetric_key: bytes, public_key: rsa.RSAPublicKey,
                               encrypted_key_path: str) -> None:
         """RSA Symmetric Key Encryption"""
